feat: Implement interactive splash screen
This change introduces an interactive splash screen.

When the user clicks the arrow icon on the splash screen, they are navigated to the main landing page, and background audio starts playing.

This was implemented by:
- Creating an AudioProvider context to manage audio playback globally.
- Updating the root layout to use the AudioProvider.
- Creating a new route for the landing page at /landing.
- Making the splash screen at the root route interactive.
