+++ app/layout.tsx
@@ -1,41 +1,50 @@
 import type React from "react"
 import type { Metadata } from "next"
-import { Inter } from "next/font/google"
+import { GeistSans } from "geist/font/sans"
+import { GeistMono } from "geist/font/mono"
+import { Analytics } from "@vercel/analytics/next"
+import { Instrument_Serif } from "next/font/google"
+import localFont from "next/font/local"
+import { Suspense } from "react"
 import "./globals.css"
-import { ThemeProvider } from "@/components/theme-provider"
+import { AudioProvider } from "./audio-context"
 
-// Configure fonts properly
-const inter = Inter({
+const instrumentSerif = Instrument_Serif({
   subsets: ["latin"],
-  variable: "--font-inter",
+  display: "swap",
+  variable: "--font-instrument-serif",
+  weight: "400",
 })
 
-// Use system fonts as fallback for Geist Sans and IBM Plex Serif
-const geistSans = {
-  variable: "--font-geist-sans",
-}
-
-const ibmPlexSerif = {
-  variable: "--font-ibm-plex-serif",
-}
+const sdGlitchRobot = localFont({
+  src: [
+    {
+      path: "https://fonts.googleapis.com/css2?family=Orbitron:wght@400;700;900&display=swap",
+      weight: "400",
+    },
+  ],
+  variable: "--font-sd-glitch-robot",
+  display: "swap",
+})
 
 export const metadata: Metadata = {
-  title: "DeployZen - AI-Powered Testing and LLM Deployment",
-  description: "AI-Powered Testing and LLM Deployment Assistant",
-    generator: 'v0.dev'
+  title: "v0 App",
+  description: "Created with v0",
+  generator: "v0.app",
 }
 
 export default function RootLayout({
   children,
-}: {
+}: Readonly<{
   children: React.ReactNode
-}) {
+}>) {
   return (
-    <html lang="en" suppressHydrationWarning>
-      <body className={`${inter.className} ${inter.variable} ${geistSans.variable} ${ibmPlexSerif.variable}`}>
-        <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
-          {children}
-        </ThemeProvider>
+    <html lang="en" className={`${instrumentSerif.variable} ${sdGlitchRobot.variable} antialiased`}>
+      <body className={`font-sans ${GeistSans.variable} ${GeistMono.variable}`}>
+        <AudioProvider>
+          <Suspense fallback={null}>{children}</Suspense>
+        </AudioProvider>
+        <Analytics />
       </body>
     </html>
   )
